requirements.txt
absl-py==1.0.0
alembic==1.7.5
astunparse==1.6.3
attrs==21.2.0
autopage==0.4.0
backports.entry-points-selectable==1.1.1
cachetools==5.0.0
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.7
click==8.0.3
cliff==3.9.0
cmaes==0.8.2
cmd2==2.3.0
cogdl==0.2.0.post1
colorama==0.4.4
colorlog==6.6.0
cycler==0.11.0
Cython==0.29.23
decorator==5.1.0
dgl==0.6.0
dgl-cu102==0.6.0rc2
distlib==0.3.3
emoji==1.6.1
filelock==3.4.0
flake8==4.0.1
flatbuffers==2.0
fonttools==4.28.1
gast==0.5.3
gensim==4.1.2
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googledrivedownloader==0.4
grave==0.0.3
greenlet==1.1.2
grpcio==1.44.0
h5py==3.6.0
huggingface-hub==0.1.2
identify==2.4.0
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
isodate==0.6.1
Jinja2==3.0.3
joblib==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==13.0.0
littleutils==0.2.2
llvmlite==0.37.0
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.0
mccabe==0.6.1
networkx==2.5
nodeenv==1.6.0
numba==0.54.1
numpy==1.22.2
oauthlib==3.2.0
ogb==1.2.6
opt-einsum==3.3.0
optuna==2.4.0
outdated==0.2.1
packaging==21.3
pandas==1.3.4
pbr==5.8.0
Pillow==8.4.0
platformdirs==2.4.0
pre-commit==2.15.0
prettytable==2.4.0
prompt-toolkit==1.0.14
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.10.0
PyInquirer==1.0.3
pyparsing==3.0.6
pyperclip==1.8.2
pyreadline3==3.3
python-dateutil==2.8.2
PyYAML==5.4.1
rdflib==6.0.2
requests==2.26.0
requests-oauthlib==1.3.1
rsa==4.8
sacremoses==0.0.46
scikit-learn==0.24.1
scipy==1.5.2
setuptools==58.0.4
setuptools-scm==6.3.2
six==1.16.0
smart-open==5.2.1
SQLAlchemy==1.4.27
stevedore==3.5.0
tabulate==0.8.7
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
texttable==1.6.4
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.0.0
tokenizers==0.10.3
toml==0.10.2
tomli==1.2.2
torch==1.8.0
torch-geometric==2.0.2
torchaudio==0.10.0+cu102
torchvision==0.9.0
tqdm==4.51.0
transformers==4.12.5
typing_extensions==4.0.0
urllib3==1.26.7
virtualenv==20.10.0
wcwidth==0.2.5
Werkzeug==2.0.3
wheel==0.37.0
wincertstore==0.2
wrapt==1.13.3
yacs==0.1.8
zipp==3.6.0